fix: share gateway port number parsing
